Arreglado errores menores

+++ SideScroller/src/gameState.h
@@ -16,10 +16,8 @@ namespace sideScroller
 	extern int screenHeight;
 	extern int score;
 	extern int gameState;
-
 	extern int verMajor;
-
-	extern float verMinor;
+	extern int verMinor;
 	enum PlayState
 	{
 		Start,
